+++ src/ExceptionInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace bupy7\xml\constructor;
+
+/**
+ * @since 2.0.0
+ */
+interface ExceptionInterface
+{
+
+}