+++ src/RuntimeException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace bupy7\xml\constructor;
+
+use RuntimeException as BaseRuntimeException;
+
+/**
+ * @since 2.0.0
+ */
+class RuntimeException extends BaseRuntimeException implements ExceptionInterface
+{
+
+}